Add more in Readme. Setup code for nested sub-dirs
