+++ AUTHORS.rst
@@ -1,0 +1,10 @@
+*******
+Authors
+*******
+
+Contributors (random)
+============================
+
+- Lynn Cherny `@arnicas <https://github.com/arnicas>`_
+- Yannick Assogba `@tafsiri <https://github.com/tafsiri>`_
+- Jim Vallandingham `@vlandham <https://github.com/vlandham>`_
