--- README
@@ -1,3 +1,0 @@
-# textkit
-
-Command line tools for text processing and analysis.
