+++ requirements.txt
@@ -1,3 +1,4 @@
+click>=6.2
 coverage>=4.0.3
 nltk>=3.1
 numpy>=1.10.2
