+++ textkit/tokenize/__init__.py
@@ -1,0 +1,2 @@
+__version__ = '0.0.1'
+__license__ = 'MIT'
