+++ textkit/tokenize/sentences.py
@@ -4,14 +4,6 @@ import click
 
 @click.command()
 @click.argument('text', type=click.File('r'))
-def text2words(text):
-    '''Tokenize text into word tokens'''
-    content = text.read()
-    [click.echo(w) for w in content.split()]
-
-
-@click.command()
-@click.argument('text', type=click.File('r'))
 def text2sentences(text):
     '''Tokenize text into sentence tokens'''
     content = text.read()
